# openapi/personalcrm_openapi_v1.yaml
openapi: 3.0.3
info:
  title: PersonalCRM API
  version: 1.0.0
  description: |
    PersonalCRM v1 API for managing contacts, interactions, introductions, reminders,
    strategy/wheel, projects, and RBAC in a cloud workspace with assistant access.

    Notes:
    - All requests are scoped to a workspace.
    - Field-level privacy is enforced server-side; assistants may receive redacted fields.
    - Use cursor-based pagination (next_cursor).
servers:
  - url: https://api.personalcrm.example.com
    description: Production (replace)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/rbac/roles:
    get:
      tags: [RBAC]
      summary: List workspace roles
      operationId: listRoles
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      responses:
        '200':
          description: Roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'
                required: [data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [RBAC]
      summary: Create a role
      operationId: createRole
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/rbac/roles/{role_id}:
    patch:
      tags: [RBAC]
      summary: Update a role
      operationId: updateRole
      parameters:
        - $ref: '#/components/parameters/RoleIdPath'
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [RBAC]
      summary: Delete a role
      operationId: deleteRole
      parameters:
        - $ref: '#/components/parameters/RoleIdPath'
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/rbac/permissions:
    get:
      tags: [RBAC]
      summary: List known permissions
      operationId: listPermissions
      responses:
        '200':
          description: Permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Permission'
                required: [data]
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/organizations:
    get:
      tags: [Organizations]
      summary: List organizations
      operationId: listOrganizations
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: q
          in: query
          schema:
            type: string
          description: Search by name
      responses:
        '200':
          description: Organizations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrganizations'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Organizations]
      summary: Create an organization
      operationId: createOrganization
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/introductions:
    get:
      tags: [Introductions]
      summary: List introductions
      operationId: listIntroductions
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/IntroductionStatus'
        - name: contact_id
          in: query
          schema:
            $ref: '#/components/schemas/UUID'
          description: Filter by a participant contact id
      responses:
        '200':
          description: Introductions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedIntroductions'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Introductions]
      summary: Create an introduction workflow
      operationId: createIntroduction
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntroductionCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Introduction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/introductions/{introduction_id}:
    get:
      tags: [Introductions]
      summary: Get an introduction
      operationId: getIntroduction
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/IntroductionIdPath'
      responses:
        '200':
          description: Introduction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Introduction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Introductions]
      summary: Update an introduction
      operationId: updateIntroduction
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/IntroductionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntroductionUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Introduction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/projects:
    get:
      tags: [Projects]
      summary: List projects
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: q
          in: query
          schema:
            type: string
          description: Search by name/description
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProjectStatus'
      responses:
        '200':
          description: Projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjects'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Projects]
      summary: Create a project
      operationId: createProject
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/projects/{project_id}:
    get:
      tags: [Projects]
      summary: Get a project
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ProjectIdPath'
      responses:
        '200':
          description: Project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Projects]
      summary: Update a project
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ProjectIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Projects]
      summary: Delete a project
      operationId: deleteProject
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
        - $ref: '#/components/parameters/ProjectIdPath'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceIdQuery:
      name: workspace_id
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
      description: Workspace scope for the request
...
    RoleIdPath:
      name: role_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
    IntroductionIdPath:
      name: introduction_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
    ProjectIdPath:
      name: project_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque pagination cursor
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    UUID:
      type: string
      format: uuid
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required: [code, message]
      required: [error]
    Visibility:
      type: string
      enum: [private, shared, limited]
      description: |
        Record visibility.
        - private: only Owner (and roles with private access)
        - shared: Owner + Assistant
        - limited: Assistant can see a redacted subset
...
    IntroductionStatus:
      type: string
      enum: [requested, approved_a, approved_b, sent, met, completed, canceled]
...
    ProjectStatus:
      type: string
      enum: [idea, active, on_hold, completed, canceled]
...
    Permission:
      type: object
      properties:
        key: { type: string }
        description: { type: string }
      required: [key]
    Role:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
        description: { type: string }
        permission_keys:
          type: array
          items: { type: string }
      required: [id, workspace_id, name, permission_keys]
    RoleCreate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        permission_keys:
          type: array
          items: { type: string }
      required: [name, permission_keys]
    RoleUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        permission_keys:
          type: array
          items: { type: string }
    Organization:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
        website: { type: string }
        industries:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, workspace_id, name, created_at, updated_at]
    OrganizationCreate:
      type: object
      properties:
        name: { type: string }
        website: { type: string }
        industries:
          type: array
          items: { type: string }
      required: [name]
...
    Introduction:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        status: { $ref: '#/components/schemas/IntroductionStatus' }

        requester_contact_id: { $ref: '#/components/schemas/UUID' }
        introducer_contact_id: { $ref: '#/components/schemas/UUID' }
        target_contact_id: { $ref: '#/components/schemas/UUID' }

        ask: { type: string }
        benefit_for_requester: { type: string }
        benefit_for_target: { type: string }

        consent_requester: { type: boolean }
        consent_target: { type: boolean }

        sent_at: { type: string, format: date-time, nullable: true }
        met_at: { type: string, format: date-time, nullable: true }
        outcome: { type: string, nullable: true }

        created_by_user_id: { $ref: '#/components/schemas/UUID' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required:
        - id
        - workspace_id
        - status
        - requester_contact_id
        - introducer_contact_id
        - target_contact_id
        - ask
        - created_at
        - updated_at
    IntroductionCreate:
      type: object
      properties:
        requester_contact_id: { $ref: '#/components/schemas/UUID' }
        introducer_contact_id: { $ref: '#/components/schemas/UUID' }
        target_contact_id: { $ref: '#/components/schemas/UUID' }
        ask: { type: string }
        benefit_for_requester: { type: string }
        benefit_for_target: { type: string }
        consent_requester: { type: boolean, default: false }
        consent_target: { type: boolean, default: false }
        status: { $ref: '#/components/schemas/IntroductionStatus' }
      required: [requester_contact_id, introducer_contact_id, target_contact_id, ask]
    IntroductionUpdate:
      type: object
      properties:
        status: { $ref: '#/components/schemas/IntroductionStatus' }
        ask: { type: string }
        benefit_for_requester: { type: string }
        benefit_for_target: { type: string }
        consent_requester: { type: boolean }
        consent_target: { type: boolean }
        sent_at: { type: string, format: date-time, nullable: true }
        met_at: { type: string, format: date-time, nullable: true }
        outcome: { type: string, nullable: true }
...
    Project:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        workspace_id: { $ref: '#/components/schemas/UUID' }
        visibility: { $ref: '#/components/schemas/Visibility' }
        name: { type: string }
        description: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        start_date: { type: string, format: date, nullable: true }
        end_date: { type: string, format: date, nullable: true }
        participant_contact_ids:
          type: array
          items: { $ref: '#/components/schemas/UUID' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, workspace_id, name, status, created_at, updated_at]
    ProjectCreate:
      type: object
      properties:
        visibility: { $ref: '#/components/schemas/Visibility' }
        name: { type: string }
        description: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        participant_contact_ids:
          type: array
          items: { $ref: '#/components/schemas/UUID' }
      required: [name]
    ProjectUpdate:
      type: object
      properties:
        visibility: { $ref: '#/components/schemas/Visibility' }
        name: { type: string }
        description: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        start_date: { type: string, format: date, nullable: true }
        end_date: { type: string, format: date, nullable: true }
        participant_contact_ids:
          type: array
          items: { $ref: '#/components/schemas/UUID' }
...
    PaginatedOrganizations:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Organization' }
        next_cursor:
          type: string
          nullable: true
      required: [data]
...
    PaginatedIntroductions:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Introduction' }
        next_cursor:
          type: string
          nullable: true
      required: [data]
...
    PaginatedProjects:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Project' }
        next_cursor:
          type: string
          nullable: true
      required: [data]
...